+++ tests/Functions/testFunctions.h
@@ -23,6 +23,8 @@ class testFunctions : public CPPUNIT_NS::TestFixture {
     CPPUNIT_TEST(testSdCircular_);
     CPPUNIT_TEST(testSdLinear_);
     CPPUNIT_TEST(testMean_);
+    CPPUNIT_TEST(testFindClosest_);
+    CPPUNIT_TEST(testFindClosest2_);
 
     CPPUNIT_TEST_SUITE_END();
 
@@ -46,6 +48,8 @@ private:
     void testSdCircular_();
     void testSdLinear_();
     void testMean_();
+    void testFindClosest_();
+    void testFindClosest2_();
     
     bool neighborExists_(const Functions::Matrix & table,
             size_t test_index, size_t neighbor_index) const;